# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "kaas-ui",
    "description": "A read-only, multi-cluster Kafka UI. Every data route is a GET; there is no mutating endpoint in this document because there is none in the router.",
    "license": {
      "name": "Apache-2.0"
    },
    "version": "0.9.0"
  },
  "paths": {
...
    "/api/environments/{env}/clusters/{id}/groups/{group}/offsets": {
      "get": {
        "tags": [
          "groups"
        ],
        "summary": "`GET /api/clusters/{id}/groups/{group}/offsets`",
        "description": "Committed offsets joined to the log ends, so lag can be classified rather\nthan subtracted. \"No commit yet\", \"empty partition\", \"caught up\" and\n\"behind by n\" are four different answers and rendering them all as `0` is\nhow a lag column becomes something nobody trusts.",
        "operationId": "offsets",
        "parameters": [
          {
            "name": "env",
            "in": "path",
            "description": "Environment id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Cluster id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "group",
            "in": "path",
            "description": "Group id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Committed offsets and lag",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_GroupOffset"
                }
              }
            }
          }
        }
      }
    },
    "/api/environments/{env}/clusters/{id}/topics": {
      "get": {
        "tags": [
          "topics"
        ],
        "summary": "`GET /api/clusters/{id}/topics`",
        "operationId": "list",
        "parameters": [
          {
            "name": "env",
            "in": "path",
            "description": "Environment id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Cluster id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Substring match",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "internal",
            "in": "query",
            "description": "Include internal topics",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "name | partitions | messages | size | underReplicated",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "asc | desc",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Page offset",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "metrics",
            "in": "query",
            "description": "Fetch message counts and sizes",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "Describe these topics instead of listing",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Topics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_TopicSummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/environments/{env}/clusters/{id}/topics/{topic}": {
      "get": {
        "tags": [
          "topics"
        ],
        "summary": "`GET /api/clusters/{id}/topics/{topic}`",
        "description": "`describe_topics` prefers `DescribeTopicPartitions` and falls back to\n`Metadata` where the newer call is unreachable. Both branches are live\nagainst the two development clusters, and **there is no branch here** —\nwhich is the claim that a partially-implemented broker is indistinguishable\nfrom an old one, made good.",
        "operationId": "detail",
        "parameters": [
          {
            "name": "env",
            "in": "path",
            "description": "Environment id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Cluster id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "topic",
            "in": "path",
            "description": "Topic name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offsets",
            "in": "query",
            "description": "Also fetch the offset range",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Topic detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_TopicDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/environments/{env}/clusters/{id}/topics/{topic}/offsets": {
      "get": {
        "tags": [
          "topics"
        ],
        "summary": "`GET /api/clusters/{id}/topics/{topic}/offsets`",
        "operationId": "offsets",
        "parameters": [
          {
            "name": "env",
            "in": "path",
            "description": "Environment id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Cluster id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "topic",
            "in": "path",
            "description": "Topic name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Offset ranges",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Partition"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Envelope_GroupOffset": {
        "type": "object",
        "description": "What every data endpoint returns.",
        "required": [
          "items",
          "errors"
        ],
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourceError"
            },
            "description": "The resources that did not. Non-empty is still `200 OK`."
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "A group's committed offset for one partition, with lag.",
              "required": [
                "topic",
                "partition",
                "lag"
              ],
              "properties": {
                "committedOffset": {
                  "type": [
                    "integer",
                    "null"
                  ],
                  "format": "int64",
                  "description": "The committed offset, where there is one."
                },
                "lag": {
                  "$ref": "#/components/schemas/Lag",
                  "description": "Lag, in its four distinguishable states."
                },
                "latestOffset": {
                  "type": [
                    "integer",
                    "null"
                  ],
                  "format": "int64",
                  "description": "The partition's next offset."
                },
                "metadata": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "description": "Commit metadata."
                },
                "partition": {
                  "type": "integer",
                  "format": "int32",
                  "description": "Partition."
                },
                "topic": {
                  "type": "string",
                  "description": "Topic."
                }
              }
            },
            "description": "The resources that answered."
          },
          "snapshotAgeMs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Age of the metadata snapshot the answer was built from, where one was\ninvolved. This is what makes \"as of 4 seconds ago\" honest.",
            "minimum": 0
          },
          "total": {
            "type": [
              "integer",
              "null"
            ],
            "description": "How many resources matched before paging or truncation.\n\nPresent only where the endpoint pages. Filtering and sorting happen on\nthe server — a fleet with a five-thousand-topic cluster in it is real —\nso the browser needs to be told what it is looking at a window into.",
            "minimum": 0
          }
        }
      },
...
      "Envelope_TopicDetail": {
        "type": "object",
        "description": "What every data endpoint returns.",
        "required": [
          "items",
          "errors"
        ],
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourceError"
            },
            "description": "The resources that did not. Non-empty is still `200 OK`."
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "A topic and its partitions.",
              "required": [
                "name",
                "internal",
                "partitions",
                "brokerIds"
              ],
              "properties": {
                "brokerIds": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "description": "The brokers holding replicas, for the placement grid's columns."
                },
                "internal": {
                  "type": "boolean",
                  "description": "Whether Kafka considers this internal."
                },
                "name": {
                  "type": "string",
                  "description": "Topic name."
                },
                "partitions": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Partition"
                  },
                  "description": "Every partition."
                },
                "topicId": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "description": "The topic uuid, where reported."
                }
              }
            },
            "description": "The resources that answered."
          },
          "snapshotAgeMs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Age of the metadata snapshot the answer was built from, where one was\ninvolved. This is what makes \"as of 4 seconds ago\" honest.",
            "minimum": 0
          },
          "total": {
            "type": [
              "integer",
              "null"
            ],
            "description": "How many resources matched before paging or truncation.\n\nPresent only where the endpoint pages. Filtering and sorting happen on\nthe server — a fleet with a five-thousand-topic cluster in it is real —\nso the browser needs to be told what it is looking at a window into.",
            "minimum": 0
          }
        }
      },
      "Envelope_TopicSummary": {
        "type": "object",
        "description": "What every data endpoint returns.",
        "required": [
          "items",
          "errors"
        ],
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourceError"
            },
            "description": "The resources that did not. Non-empty is still `200 OK`."
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "One row of the topic list.",
              "required": [
                "name",
                "internal",
                "partitionCount",
                "replicationFactor",
                "offlinePartitionCount",
                "underReplicatedPartitionCount"
              ],
              "properties": {
                "internal": {
                  "type": "boolean",
                  "description": "Whether Kafka considers this internal."
                },
                "logicalBytes": {
                  "type": [
                    "integer",
                    "null"
                  ],
                  "format": "int64",
                  "description": "Bytes on disk for one copy, when metrics were asked for."
                },
                "messageCount": {
                  "type": [
                    "integer",
                    "null"
                  ],
                  "format": "int64",
                  "description": "Records retained across every partition, when metrics were asked for.\n\n`latest - earliest` summed, which is what is *retained* rather than\nwhat was ever written — the same distinction the partition table draws.\n`None` when the metric was not requested, and also when any one\npartition failed to answer: a sum missing a partition is a smaller\nnumber, not a marked one, and nothing in the column would say so."
                },
                "name": {
                  "type": "string",
                  "description": "Topic name."
                },
                "offlinePartitionCount": {
                  "type": "integer",
                  "description": "Partitions with no leader or an offline replica.",
                  "minimum": 0
                },
                "partitionCount": {
                  "type": "integer",
                  "description": "Partition count.",
                  "minimum": 0
                },
                "replicatedBytes": {
                  "type": [
                    "integer",
                    "null"
                  ],
                  "format": "int64",
                  "description": "Bytes on disk across replicas."
                },
                "replicationFactor": {
                  "type": "integer",
                  "description": "The smallest replica count across partitions, which is what anyone\nmeans by \"replication factor\".",
                  "minimum": 0
                },
                "topicId": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "description": "The topic uuid, or `None` on a cluster that does not report one.\n\nAbsent rather than an empty uuid: a column of zeroes reads as missing\ndata, and the honest render is no column at all."
                },
                "underReplicatedPartitionCount": {
                  "type": "integer",
                  "description": "Partitions whose ISR is short.",
                  "minimum": 0
                }
              }
            },
            "description": "The resources that answered."
          },
          "snapshotAgeMs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Age of the metadata snapshot the answer was built from, where one was\ninvolved. This is what makes \"as of 4 seconds ago\" honest.",
            "minimum": 0
          },
          "total": {
            "type": [
              "integer",
              "null"
            ],
            "description": "How many resources matched before paging or truncation.\n\nPresent only where the endpoint pages. Filtering and sorting happen on\nthe server — a fleet with a five-thousand-topic cluster in it is real —\nso the browser needs to be told what it is looking at a window into.",
            "minimum": 0
          }
        }
      },
...
    }
  },
...
}